requirements.txt
openai~=1.8.0
PyMuPDF~=1.23.15
python-dotenv~=1.0.0
pillow~=10.2.0
loguru~=0.7.2
